Refactoring nearbyservice to singleton
