+++ server/src/main/java/com/amos/server/ConnectToClientActivity.java
@@ -17,24 +17,8 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.android.gms.nearby.Nearby;
-import com.google.android.gms.nearby.connection.ConnectionInfo;
-import com.google.android.gms.nearby.connection.ConnectionLifecycleCallback;
-import com.google.android.gms.nearby.connection.ConnectionResolution;
-import com.google.android.gms.nearby.connection.ConnectionsClient;
-import com.google.android.gms.nearby.connection.ConnectionsStatusCodes;
-import com.google.android.gms.nearby.connection.DiscoveredEndpointInfo;
-import com.google.android.gms.nearby.connection.DiscoveryOptions;
-import com.google.android.gms.nearby.connection.EndpointDiscoveryCallback;
-import com.google.android.gms.nearby.connection.Payload;
-import com.google.android.gms.nearby.connection.PayloadCallback;
-import com.google.android.gms.nearby.connection.PayloadTransferUpdate;
-import com.google.android.gms.nearby.connection.Strategy;
-
-import java.security.SecureRandom;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
+import com.amos.server.nearby.ConnectCallback;
+import com.amos.server.nearby.ServerConnection;
 
 public class ConnectToClientActivity extends Activity {
 
@@ -49,221 +33,58 @@ public class ConnectToClientActivity extends Activity {
 
     private static final int REQUEST_CODE_REQUIRED_PERMISSIONS = 1;
 
-    /** 1-to-1 since a device will be connected to only one other device at most. */
-    private static final Strategy STRATEGY = Strategy.P2P_POINT_TO_POINT;
-
-    /** Connection manager for the connection to FlyInn clients. */
-    protected ConnectionsClient connectionsClient;
-
-    private final String clientName = generateName(5);
-    private String clientID;
-    private String clientNameNow;
-
-    /** Toast to publish user notifications */
-    private Toast mToast;
-
-    /** List of all discovered servers by name, continuously updated. */
-    private List<String> clients = new ArrayList<>();
-
-    /** Maps server names to their nearby connection IDs. */
-    private HashMap<String, String> clientNamesToIDs = new HashMap<>();
-
-    /** Maps server IDs to their nearby connection names. */
-    private HashMap<String, String> clientIDsToNames = new HashMap<>();
+    private ServerConnection connection = ServerConnection.getInstance();
 
     /** Tag for logging purposes. */
     private static final String TAG = "ClientNearbyConnection";
-
-    /**
-     * Obtain data from clientID/clientName and data transfer information via this handle.
-     */
-    private final PayloadCallback payloadCallback =
-            new PayloadCallback() {
-                @Override
-                public void onPayloadReceived(String endpointId, Payload payload) {
-                    Log.d(TAG, "Payload received");
-                }
-
-                @Override
-                public void onPayloadTransferUpdate(String endpointId, PayloadTransferUpdate update) {
-                    Log.d(TAG, "Payload transfer update");
-                }
-            };
-
-
-    /**
-     * Handling of discovered endpoints (servers). Adds new endpoints to servers data maps/list,
-     * and removes lost endpoints.
-     */
-    private final EndpointDiscoveryCallback endpointDiscoveryCallback =
-            new EndpointDiscoveryCallback() {
-                @Override
-                public void onEndpointFound(String endpointId, DiscoveredEndpointInfo info) {
-                    // discovered a server, add to data maps
-                    String endpointName = info.getEndpointName();
-
-                    if (!(clientIDsToNames.containsKey(endpointId)
-                            || clientNamesToIDs.containsKey(endpointName))) {
-                        clients.add(endpointName);
-                        clientNamesToIDs.put(endpointName, endpointId);
-                        clientIDsToNames.put(endpointId, endpointName);
-                        Log.i(TAG, clientName + " discovered endpoint " + endpointId + " with name " + endpointName);
-
-                    } else {
-                        // this should not happen
-                        while (clients.remove(endpointName)) {}
-                        clients.add(endpointName);
-                        clientIDsToNames.put(endpointId, endpointName);
-                        clientNamesToIDs.put(endpointName, endpointId);
-                        Log.i(TAG, clientName + " rediscovered endpoint " + endpointId + " with name " + endpointName);
-                    }
-                }
-
-                @Override
-                public void onEndpointLost(String endpointId) {
-                    // previously discovered server is no longer reachable, remove from data maps
-                    String lostEndpointName = clientIDsToNames.get(endpointId);
-                    clientIDsToNames.remove(endpointId);
-                    clientNamesToIDs.remove(lostEndpointName);
-                    Log.i(TAG, clientName + " lost discovered endpoint " + endpointId);
-                }
-            };
-
-
-    /**
-     * Callbacks for connections to other devices.
-     * Includes token authentication and connection handling.
-     */
-    private final ConnectionLifecycleCallback connectionLifecycleCallback =
-            new ConnectionLifecycleCallback() {
-                @Override
-                public void onConnectionInitiated(String endpointId, ConnectionInfo connectionInfo) {
-                    Log.i(TAG, "Connection initiated to " + endpointId);
-
-                    if (endpointId.equals(clientID)) {
-                        connectionsClient.acceptConnection(endpointId, payloadCallback);
-                    } else {
-                        // initiated connection is not with server selected by user
-                        connectionsClient.rejectConnection(endpointId);
-                        Log.i(TAG, "Connection rejected to non-selected server "
-                                + endpointId);
-                    }
-                }
-
-                @Override
-                public void onConnectionResult(String endpointId, ConnectionResolution result) {
-                    switch (result.getStatus().getStatusCode()) {
-
-                        case ConnectionsStatusCodes.STATUS_OK:
-                            // successful connection with server
-                            Log.i(TAG, "Connected with " + endpointId);
-                            mToast.setText(R.string.nearby_connection_success);
-                            mToast.show();
-                            connectedToApp();
-                            break;
-
-                        case ConnectionsStatusCodes.STATUS_CONNECTION_REJECTED:
-                            // connection was rejected by one side (or both)
-                            Log.i(TAG, "Connection rejected with " + endpointId);
-                            mToast.setText(R.string.nearby_connection_rejected);
-                            mToast.show();
-                            clientNameNow = null;
-                            clientID = null;
-                            break;
-
-                        case ConnectionsStatusCodes.STATUS_ERROR:
-                            // connection was lost
-                            Log.w(TAG, "Connection lost: " + endpointId);
-                            mToast.setText(R.string.nearby_connection_error);
-                            mToast.show();
-                            clientNameNow = null;
-                            clientID = null;
-                            break;
-
-                        default:
-                            // unknown status code. we shouldn't be here
-                            Log.e(TAG, "Unknown error when attempting to connect with "
-                                    + endpointId);
-                            mToast.setText(R.string.nearby_connection_error);
-                            mToast.show();
-                            clientNameNow = null;
-                            clientID = null;
-                    }
-                }
-
-                @Override
-                public void onDisconnected(String endpointId) {
-                    // disconnected from server
-                    Log.i(TAG, "Disconnected from " + endpointId);
-                    mToast.setText(R.string.nearby_disconnected);
-                    mToast.show();
-                    clearServerData();
-                    finish();
-                }
-            };
-
-
-
-    private void requestConnectionWithClient(String code){
-
-        String searchedClientName = null;
-        for (String clientName : clients){
-            if(clientName.endsWith(code))
-            {
-                searchedClientName = clientName;
-                break;
-            }
-        }
-
-        if (searchedClientName == null)
-        {
-            Toast.makeText(this,"The code given was not found. Please try again",Toast.LENGTH_LONG).show();
-            return;
-        }
-
-
-        String endpoint = clientNamesToIDs.get(searchedClientName);
-        clientID = endpoint;
-
-
-        connectionsClient.requestConnection(searchedClientName,endpoint,connectionLifecycleCallback);
-
-
-
-    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_connect_to_client);
         EditText text = findViewById(R.id.connect_editText);
-        text.setOnEditorActionListener(new TextView.OnEditorActionListener() {
-            @Override
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                if (actionId == EditorInfo.IME_ACTION_DONE) {
-                    String name = v.getText().toString(); // Get the String
-                    requestConnectionWithClient(name);
-                    return true;
-                }
-                return false;
-            }
-        });
 
+        checkPermissions();
+        // Ensure survival for life of entire application
+        connection.init(getApplicationContext());
+        connection.discover();
+        text.setOnEditorActionListener(
+                (TextView v, int actionId, KeyEvent event) -> {
+                    if (actionId == EditorInfo.IME_ACTION_DONE) {
+                        String name = v.getText().toString(); // Get the String
+                        toast("Connecting to " + name);
+                        connection.connectTo(name, new ConnectCallback() {
+                            @Override
+                            public void success() {
+                                connectionSuccess(name);
+                            }
 
-        if (!hasPermissions(this, REQUIRED_PERMISSIONS) &&
-                Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            requestPermissions(REQUIRED_PERMISSIONS, REQUEST_CODE_REQUIRED_PERMISSIONS);
-        } else {
-            Log.w(TAG, "Could not check permissions due to version");
-        }
+                            @Override
+                            public void failure() {
+                                connectionFailure(name);
+                            }
+                        });
+                        return true;
+                    }
+                    return false;
+                });
+    }
 
-        connectionsClient = Nearby.getConnectionsClient(this);
-        mToast = Toast.makeText(this, "", Toast.LENGTH_SHORT);
+    /**
+     * Switch to activity handling an established connection.
+     */
+    private void connectionSuccess(String name) {
+        toast(String.format("Successfully connected to %s", name));
+        Intent intent = new Intent(this, ConnectionSetupServerActivity.class);
+        startActivity(intent);
+    }
 
-
-        startDiscovering();
-
-
+    /**
+     * Handle failure to connection to given code
+     * @param name
+     */
+    private void connectionFailure(String name) {
+        toast(String.format("Failed to connect to %s", name));
     }
 
     /**
@@ -274,12 +95,7 @@ public class ConnectToClientActivity extends Activity {
         super.onStart();
 
         // user may have changed permissions
-        if (!hasPermissions(this, REQUIRED_PERMISSIONS) &&
-                Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            requestPermissions(REQUIRED_PERMISSIONS, REQUEST_CODE_REQUIRED_PERMISSIONS);
-        } else {
-            Log.w(TAG, "Could not check permissions due to version");
-        }
+        checkPermissions();
     }
 
     /**
@@ -288,64 +104,21 @@ public class ConnectToClientActivity extends Activity {
      */
     @Override
     protected void onDestroy() {
-        connectionsClient.stopDiscovery();
-        connectionsClient.stopAllEndpoints();
-        clearServerData();
         super.onDestroy();
+        connection.abort();
     }
 
-    /**
-     * Clears all servers map data as well as serverName/serverID and starts discovery
-     */
-    private void startDiscovering() {
-        clearServerData();
-
-        DiscoveryOptions discoveryOptions =
-                new DiscoveryOptions.Builder().setStrategy(STRATEGY).build();
-
-        connectionsClient.startDiscovery("com.amos.server", endpointDiscoveryCallback,
-                discoveryOptions)
-                .addOnSuccessListener( (Void unused) -> {
-                    // started searching for servers successfully
-                    Log.i(TAG, "Discovering connections on " + clientName);
-                    mToast.setText(R.string.nearby_discovering_success);
-                    mToast.show();
-                })
-                .addOnFailureListener( (Exception e) -> {
-                    // unable to start discovery
-                    Log.e(TAG, "Unable to start discovery on " + clientName);
-                    mToast.setText(R.string.nearby_discovering_error);
-                    mToast.show();
-                    finish();
-                });
-    }
-
-    /**
-     * Clears serverName/serverID and all server data maps as well as the servers list
-     */
-    private void clearServerData() {
-        clients.clear();
-        clientIDsToNames.clear();
-        clientNamesToIDs.clear();
-        clientID = null;
-        clientNameNow = null;
-    }
-
-    /**
-     * Handle established connection with app.
-     *
-     * Clears servers data maps, stops discovery of new servers and adds close connection button
-     */
-    private void connectedToApp() {
-        connectionsClient.stopDiscovery();
-        clients.clear();
-        clientNamesToIDs.clear();
-        clientIDsToNames.clear();
-
-        // Start setting up a connection with the other side
-        Intent intent = new Intent(this, ConnectionSetupServerActivity.class);
-        intent.putExtra("endpointId", clientID);
-        startActivity(intent);
+    private void checkPermissions() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            if (!hasPermissions(this, REQUIRED_PERMISSIONS)) {
+                requestPermissions(REQUIRED_PERMISSIONS, REQUEST_CODE_REQUIRED_PERMISSIONS);
+            } else {
+                Log.d(TAG, "Permissions are ok.");
+            }
+        } else {
+            Log.w(TAG, "Could not check permissions due to version");
+            toast("Could not check permissions due to version");
+        }
     }
 
     /**
@@ -362,6 +135,11 @@ public class ConnectToClientActivity extends Activity {
             }
         }
         return true;
+    }
+
+    private void toast(String message) {
+        Toast toast = Toast.makeText(this, message, Toast.LENGTH_SHORT);
+        toast.show();
     }
 
     /**
@@ -384,31 +162,11 @@ public class ConnectToClientActivity extends Activity {
             if (grantResult == PackageManager.PERMISSION_DENIED) {
                 Log.w(TAG, "Permissions necessary for " +
                         "Nearby Connection were not granted.");
-                mToast.setText(R.string.nearby_missing_permissions);
-                mToast.show();
+                toast("Permissions are missing for nearby");
                 finish();
             }
         }
         recreate();
-    }
-
-    /**
-     * Generates a name for the server.
-     * @return The server name, consisting of the build model + a random string
-     */
-    // TODO Define better name system?
-    private String generateName(int appendixLength){
-        String AB = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-        SecureRandom rnd = new SecureRandom();
-
-        StringBuilder sb = new StringBuilder(appendixLength);
-        for (int i = 0; i < appendixLength; i++) {
-            sb.append(AB.charAt(rnd.nextInt(AB.length())));
-        }
-
-        String name = Build.MODEL + "_" + sb.toString();
-        Log.i(TAG, "Current name is: " + name);
-        return name;
     }
 
 }
