+++ server/src/main/java/com/amos/server/nearby/ConnectCallback.java
@@ -1,0 +1,6 @@
+package com.amos.server.nearby;
+
+public interface ConnectCallback {
+    void success();
+    void failure();
+}
